# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "DriverSummary" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "DOCS_SUBMITTED", "TRAINING", "APPROVED", "REJECTED", "SUSPENDED" ],
            "type" : "string"
          },
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageResponseDriverSummary" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/DriverSummary"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Problem" : {
        "additionalProperties" : true,
        "description" : "RFC 7807 problem details. Extra members (e.g. `missing`, `errors`) depend on `code`.",
        "properties" : {
          "code" : {
            "description" : "Stable error code for clients, e.g. `driver.documents-missing`",
            "example" : "driver.documents-missing"
          },
          "detail" : {
            "example" : "Some required documents have not been uploaded."
          },
          "instance" : { },
          "status" : {
            "example" : 409
          },
          "title" : {
            "example" : "Conflict"
          },
          "type" : {
            "example" : "about:blank"
          }
        }
      },
...
      "ReasonRequest" : {
        "properties" : {
          "reason" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RegisterDriverRequest" : {
        "properties" : {
          "bankAccountName" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "bankAccountNumber" : {
            "type" : "string"
          },
          "bankName" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "fullName" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "nationalId" : {
            "type" : "string"
          },
          "phone" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "fullName", "phone" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from /api/v1/auth/otp/verify or /api/v1/auth/refresh",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Scheduled & contracted mobility platform for Phnom Penh.\n\n* **Authentication:** phone OTP (`/api/v1/auth/otp/request` → `/otp/verify`) returns a 15-minute JWT access token. Send it as `Authorization: Bearer <token>`. Refresh with `/api/v1/auth/refresh`.\n* **Retries:** mutating endpoints accept an `Idempotency-Key` header; a retry with the same key and body replays the original response (`Idempotent-Replayed: true`).\n* **Errors:** RFC 7807 `application/problem+json` with a stable machine-readable `code`; `title` and `detail` are localized from `Accept-Language` (`km` default, `en`).\n* **Time:** timestamps are UTC (ISO-8601); business dates are Asia/Phnom_Penh.",
    "title" : "Mobility API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/admin/drivers" : {
      "get" : {
        "description" : "Newest first. Filter by `status`; `size` is 1–100.\n\n**Roles:** ADMIN, DISPATCHER, SUPPORT, SAFETY_OFFICER",
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "PENDING", "DOCS_SUBMITTED", "TRAINING", "APPROVED", "REJECTED", "SUSPENDED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDriverSummary"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "List drivers",
        "tags" : [ "Admin · Drivers" ],
        "x-required-roles" : [ "ADMIN", "DISPATCHER", "SUPPORT", "SAFETY_OFFICER" ]
      },
      "post" : {
        "description" : "Creates the driver (PENDING) and their login (DRIVER role) for the phone number; an existing passenger account is reused. National ID and bank account are encrypted at rest.\n\n**Roles:** ADMIN",
        "operationId" : "register_1",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterDriverRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverDetail"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: invalid state transition, duplicate or concurrent update (`code` explains which)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Register a driver",
        "tags" : [ "Admin · Drivers" ],
        "x-required-roles" : [ "ADMIN" ]
      }
    },
    "/api/v1/admin/drivers/{driverId}" : {
      "get" : {
        "description" : "Includes current documents, vehicle, `readiness` (what blocks approval) and status history.\n\n**Roles:** ADMIN, DISPATCHER, SUPPORT, SAFETY_OFFICER",
        "operationId" : "get_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverDetail"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Driver detail",
        "tags" : [ "Admin · Drivers" ],
        "x-required-roles" : [ "ADMIN", "DISPATCHER", "SUPPORT", "SAFETY_OFFICER" ]
      }
    },
    "/api/v1/admin/drivers/{driverId}/approve" : {
      "post" : {
        "description" : "Also requires an assigned vehicle with approved VEHICLE_REGISTRATION and VEHICLE_INSURANCE. The driver can then be dispatched.\n\n**Roles:** ADMIN",
        "operationId" : "approve",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverDetail"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Not found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: invalid state transition, duplicate or concurrent update (`code` explains which)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "TRAINING → APPROVED",
        "tags" : [ "Admin · Drivers" ],
        "x-required-roles" : [ "ADMIN" ]
      }
    },
    "/api/v1/admin/drivers/{driverId}/documents/{documentId}/approve" : {
      "post" : {
        "description" : "Records reviewer and time. Expired documents cannot be approved.\n\n**Roles:** ADMIN",
        "operationId" : "approveDocument",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "documentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Not found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: invalid state transition, duplicate or concurrent update (`code` explains which)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Approve a document",
        "tags" : [ "Admin · Drivers" ],
        "x-required-roles" : [ "ADMIN" ]
      }
    },
    "/api/v1/admin/drivers/{driverId}/documents/{documentId}/content" : {
      "get" : {
        "description" : "Decrypted file. Every access is written to the audit log.\n\n**Roles:** ADMIN, SAFETY_OFFICER",
        "operationId" : "documentContent",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "documentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              },
              "image/jpeg" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              },
              "image/png" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              }
            },
            "description" : "The file"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Download a document file",
        "tags" : [ "Admin · Drivers" ],
        "x-required-roles" : [ "ADMIN", "SAFETY_OFFICER" ]
      }
    },
    "/api/v1/admin/drivers/{driverId}/documents/{documentId}/reject" : {
      "post" : {
        "description" : "`reason` (≥ 10 characters) is shown to the driver, who must upload a new one.\n\n**Roles:** ADMIN",
        "operationId" : "rejectDocument",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "documentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Not found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: invalid state transition, duplicate or concurrent update (`code` explains which)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Reject a document",
        "tags" : [ "Admin · Drivers" ],
        "x-required-roles" : [ "ADMIN" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "This server",
    "url" : "/"
  } ],
...
}